fix(cross.yml): only run on the release branch (#213)
Part of https://github.com/ooni/probe/issues/1335.

Motivation: we want all workflows to be green only when we are
approaching a release. It's fine if some less core tests are
failing during the development process. We have daily builds anyway
so we know of new breakages the day after, which is OK.