fix: remove publish-android.bash
Bintray is gone, so we cannot publish an Android release.

By removing this part of the Android build process we are going to
ensure Android CI builds complete successfully.

This is a failed build:

https://github.com/ooni/probe-cli/pull/313/checks?check_run_id=2438302064#step:8:7

As you can see, the failure here is that the file name is not the
correct file name anymore. I could fix that, but still the code would
not be able to upload anyway. Bintray is shutting down.

So, I have chosen to just remove the upload script.

Part of https://github.com/ooni/probe/issues/1439.