fix(build-ios.bash): github actions does not have git
So we need to fake the version number. It's fine for it to be
the SHA as long as we're just making a temporary build.

Since bintray is gone, there is no point for now in making more
than just a temporary build at github actions.

Part of https://github.com/ooni/probe/issues/1439