fix(make): smart update strategy for sdk/go
Rather than recording the sha256sums and downloading
the binary ourselves from golang.org/dl, do what VSCode
would do: download the binary using an existing golang
installation with `go get golang.org/dl/go$version`.

Then run `$GOPATH/bin/go$version download` to download
the SDK in the `$HOME/sdk/go$version` folder.

In addition to simplifying the code, this technique
will also reduce duplication with VSCode because we'll
download and use a single copy of the SDK. This will
also be more consistent in terms of binaries in the
`$GOPATH/bin` folder as well as in tests results.