feat(torsf): experiment that bootstraps tor using snowflake
The current implementation assumes the user has already installed tor
on the current system. If tor is not present, the experiment fails.

This is meant to be a first version of this experiment.

We are going to add more functionality in subsequent revisions of
this experiment, once we've collected more feeback.

Reference issue: https://github.com/ooni/probe/issues/1565.