fix(netxlite): close quic packetconn
Noticed when working on https://github.com/ooni/probe/issues/1505.

Justification for this diff:

1. [https://github.com/lucas-clemente/quic-go/blob/v0.21.1/client.go#L153](
DialEarlyContext calls dialContext with the last argument set to false);

2. [https://github.com/lucas-clemente/quic-go/blob/v0.21.1/client.go#L187](
the semantics of the last argument is whether we own the connection);

3. [https://github.com/lucas-clemente/quic-go/blob/v0.21.1/client.go#L269](
this value is propagated to the client data structure);

4. [https://github.com/lucas-clemente/quic-go/blob/v0.21.1/client.go#L302](
client.dial) runs the session in a background goroutine and only destroys
the `packetHandlers` when the connection is owned;

5. [https://github.com/lucas-clemente/quic-go/blob/v0.21.1/packet_handler_map.go#L293](
packetHandlerMap.Destroy) closes the underlying PacketConn.

6. also, the documentation clearly states that when you use
`DialEarlyContext` you can use the same packet conn multiple
times, so it does not take ownership.